requirements.txt
Authlib==1.3.0
black==23.10.0
celery==5.2.7
flake8==6.1.0
Flask==2.2.5
flask-babel==3.1.0
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.2.1
gunicorn==20.1.0
ics==0.7.2
isort==5.12.0
marshmallow-sqlalchemy==0.29.0
psycopg2-binary==2.9.6
pytest==7.4.2
python-dotenv==1.0.0
redis==4.5.4
requests==2.27.1
SQLAlchemy==2.0.9